Hubungkan pengguna dengan pengurus

+++ database/migrations/2022_11_09_143029_create_pengurus_table.php
@@ -19,8 +19,8 @@ class CreatePengurusTable extends Migration
             $table->string('nama', 255);
             $table->string('gelar_depan', 255)->nullable();
             $table->string('gelar_belakang', 255)->nullable();
-            $table->integer('nip')->nullable()->unique();
-            $table->integer('nik')->unique();
+            $table->bigInteger('nip')->nullable()->unique();
+            $table->bigInteger('nik')->unique();
             $table->boolean('status')->default(1); // 0: Tidak Aktif; 1: aktif
             $table->string('foto', 255)->nullable();
             $table->string('tempat_lahir', 255);
